+++ gadgets/mri_core/generic_recon_gadgets/GenericReconImageArrayScalingGadget.cpp
@@ -62,7 +62,7 @@ namespace Gadgetron {
 
         mrd::ImageArray* recon_res_ = m1->getObjectPtr();
 
-        size_t encoding = (size_t)std::get<long>(recon_res_->meta[0]["encoding"].front());
+        size_t encoding = (size_t)std::get<long long>(recon_res_->meta[0]["encoding"].front());
         GADGET_CHECK_RETURN(encoding<num_encoding_spaces_, GADGET_FAIL);
 
         std::string dataRole = std::get<std::string>(recon_res_->meta[0][GADGETRON_DATA_ROLE].front());
